+++ packages/csv/src/lookback.ts
@@ -4,19 +4,22 @@ import { CsvBlockOptions, getOptions } from "./options";
 import { parseCsv } from "./parseCsv";
 
 export const csvBlockLookBack = (
-  userOptions?: Partial<CsvBlockOptions>,
+  options: CsvBlockOptions,
 ): LookBackFunction => {
-  const options = getOptions(userOptions);
+  const completeOptions = getOptions(options);
   const {
     allIndexesVisible,
     visibleIndexes: visibleIndexesOption,
     delimiter,
-  } = options;
+  } = completeOptions;
 
   return ({ output, isComplete, visibleTextLengthTarget }) => {
-    const array = parseCsv(removeStartEndChars(output, options), options);
+    const array = parseCsv(
+      removeStartEndChars(output, completeOptions),
+      completeOptions,
+    );
     const visibleIndexes = allIndexesVisible
-      ? array.map((_, i) => i)
+      ? array.map((_, i) => i).slice(1) // Skip index 0 as it is the type
       : visibleIndexesOption;
 
     let charsRemaining = visibleTextLengthTarget;