+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @llm-ui/shared
 
+## 0.13.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @llm-ui/react@0.13.0
+
+## 0.12.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @llm-ui/react@0.12.0
+
+## 0.11.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @llm-ui/react@0.11.1
+
 ## 0.11.0
 
 ### Minor Changes